Update Contract Creation
Hi, I think the term zero address can be extremely confusing as sending a transaction to the actual zero address ("address(0)" in Solidity) does not create a new contract (the zero address is a normal address from the eyes of the evm). 

From the solidity docs: 

"If the target account is not set (the transaction does not have a recipient or the recipient is set to null), the transaction creates a new contract."
Link: https://docs.soliditylang.org/en/v0.8.19/introduction-to-smart-contracts.html#index-8

I think this makes it clearer for new developers as a lot of developers in the ecosystem mistakenly think that creating a contract is by sending the bytecode to the address(0).